Define _model_checks in AnalyzerBase init ...
...and add it to all methods that add model checks before calling `super().__init__`. 

I hope this makes the code a little more readable. 
* inside of analyzer child classes, functions like `_add_model_softmax_check` call `_add_model_check` from AnalyzerBase
* previously, this would check `hasattr(self, "_model_checks")` and add an empty list `self._model_checks = []` if necessary.
* after this, the check is added in the child class
* the child class calls `super().__init__`, which runs the checks through `_do_model_checks`

This basic mechanism has not changed, but defining the attribute _model_checks in the init should make this behaviour a bit more transparent.

The softmax check in `InputTimesGradient` was removed as it should already be added in the following `super().__init__` of `Gradient`.